+++ database/factories/UserFactory.php
@@ -25,11 +25,13 @@ class UserFactory extends Factory
     public function definition(): array
     {
         return [
+            'company_id' => null,
             'name' => fake()->name(),
             'email' => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
             'password' => static::$password ??= Hash::make('password'),
             'remember_token' => Str::random(10),
+            'status' => 'active',
         ];
     }